refactor(openmetrics): split formatter into formatter/ modules + mirror tests
