+++ packages/xo-server-openmetrics/src/openmetric-formatter.test.mts
@@ -1,2476 +1,25 @@
 /**
  * Tests for OpenMetrics Formatter Module
+ *
+ * Covers the XOSTOR formatters that still live in `openmetric-formatter.mts`
+ * and the `indexSrUuidTruncations` helper. The RRD/entity/primitive/definition
+ * formatters are tested under `./formatter/*.test.mts`.
  */
 
 import assert from 'node:assert/strict'
 import { describe, it } from 'node:test'
 
 import {
-  HOST_METRICS,
-  VM_METRICS,
-  findMetricDefinition,
-  transformMetric,
   formatToOpenMetrics,
-  formatAllPoolsToOpenMetrics,
-  formatHostStatusMetrics,
-  formatHostUptimeMetrics,
-  formatSrMetrics,
-  formatVdiMetrics,
-  formatVmStatusMetrics,
-  formatVmUptimeMetrics,
   formatXostorAlarmsMetrics,
   formatXostorClusterMetrics,
   formatXostorSmartMetrics,
   formatXostorUpdatesMetrics,
-  type FormattedMetric,
-  type LabelContext,
 } from './openmetric-formatter.mjs'
 
 import { indexSrUuidTruncations } from './index.mjs'
 
-import type {
-  HostStatusItem,
-  SrDataItem,
-  VdiDataItem,
-  VmStatusItem,
-  XostorAlarmsItem,
-  XostorClusterItem,
-  XostorSmartHost,
-  XostorUpdateItem,
-} from './index.mjs'
-
-import type { ParsedMetric, ParsedRrdData } from './rrd-parser.mjs'
-
-// ============================================================================
-// Metric Definitions Tests
-// ============================================================================
-
-describe('HOST_METRICS', () => {
-  it('should have metric definitions', () => {
-    assert.ok(HOST_METRICS.length > 0)
-  })
-
-  it('should include loadavg metric', () => {
-    const loadMetric = HOST_METRICS.find(m => m.openMetricName === 'host_load_average')
-    assert.ok(loadMetric)
-    assert.equal(loadMetric.type, 'gauge')
-  })
-
-  it('should include memory metrics with transformations', () => {
-    const memFree = HOST_METRICS.find(m => m.openMetricName === 'host_memory_free_bytes')
-    assert.ok(memFree)
-    assert.ok(memFree.transformValue)
-    // memory_free_kib * 1024 = bytes
-    assert.equal(memFree.transformValue!(1024), 1024 * 1024)
-  })
-
-  it('should include CPU core metric with label extraction', () => {
-    const cpuCore = HOST_METRICS.find(m => m.openMetricName === 'host_cpu_core_usage')
-    assert.ok(cpuCore)
-    assert.ok(cpuCore.extractLabels)
-
-    // Test regex and label extraction
-    const regex = cpuCore.test as RegExp
-    const match = regex.exec('cpu0')
-    assert.ok(match)
-    assert.deepEqual(cpuCore.extractLabels!(match), { core: '0' })
-  })
-
-  it('should include network metrics', () => {
-    const netRx = HOST_METRICS.find(m => m.openMetricName === 'host_network_receive_bytes_total')
-    assert.ok(netRx)
-    assert.equal(netRx.type, 'counter')
-  })
-
-  it('should include disk throughput metrics with MB to bytes transformation', () => {
-    const throughput = HOST_METRICS.find(m => m.openMetricName === 'host_disk_throughput_read_bytes')
-    assert.ok(throughput)
-    assert.ok(throughput.transformValue)
-    // 1 MB = 2^20 bytes
-    assert.equal(throughput.transformValue!(1), Math.pow(2, 20))
-  })
-
-  it('should include latency metrics with µs to seconds transformation', () => {
-    const latency = HOST_METRICS.find(m => m.openMetricName === 'host_disk_read_latency_seconds')
-    assert.ok(latency)
-    assert.ok(latency.transformValue)
-    // 1000000 µs = 1 second
-    assert.equal(latency.transformValue!(1000000), 1)
-  })
-})
-
-describe('VM_METRICS', () => {
-  it('should have metric definitions', () => {
-    assert.ok(VM_METRICS.length > 0)
-  })
-
-  it('should include memory metrics', () => {
-    const memMetric = VM_METRICS.find(m => m.openMetricName === 'vm_memory_bytes')
-    assert.ok(memMetric)
-  })
-
-  it('should include CPU usage metric', () => {
-    const cpuUsage = VM_METRICS.find(m => m.openMetricName === 'vm_cpu_usage')
-    assert.ok(cpuUsage)
-    assert.equal(cpuUsage.type, 'gauge')
-  })
-
-  it('should include VIF metrics with label extraction', () => {
-    const vifRx = VM_METRICS.find(m => m.openMetricName === 'vm_network_receive_bytes_total')
-    assert.ok(vifRx)
-    assert.ok(vifRx.extractLabels)
-
-    const regex = vifRx.test as RegExp
-    const match = regex.exec('vif_0_rx')
-    assert.ok(match)
-    assert.deepEqual(vifRx.extractLabels!(match), { vif: '0' })
-  })
-
-  it('should include VBD metrics with device label', () => {
-    const vbdRead = VM_METRICS.find(m => m.openMetricName === 'vm_disk_read_bytes_total')
-    assert.ok(vbdRead)
-    assert.ok(vbdRead.extractLabels)
-
-    const regex = vbdRead.test as RegExp
-    const match = regex.exec('vbd_xvda_read')
-    assert.ok(match)
-    assert.deepEqual(vbdRead.extractLabels!(match), { device: 'xvda' })
-  })
-
-  it('should include runstate metrics', () => {
-    const runstate = VM_METRICS.find(m => m.openMetricName === 'vm_runstate_fullrun')
-    assert.ok(runstate)
-    assert.equal(runstate.type, 'gauge')
-  })
-
-  it('should include VBD throughput metrics with transformation', () => {
-    const throughputRead = VM_METRICS.find(m => m.openMetricName === 'vm_disk_throughput_read_bytes')
-    assert.ok(throughputRead)
-    assert.equal(throughputRead.type, 'gauge')
-    assert.ok(throughputRead.transformValue)
-    // 1.5 MiB/s * 2^20 = 1572864 bytes/s
-    assert.equal(throughputRead.transformValue!(1.5), 1.5 * 2 ** 20)
-
-    const regex = throughputRead.test as RegExp
-    const match = regex.exec('vbd_xvda_io_throughput_read')
-    assert.ok(match)
-    assert.deepEqual(throughputRead.extractLabels!(match), { device: 'xvda' })
-  })
-
-  it('should include VBD throughput write and total metrics', () => {
-    const throughputWrite = VM_METRICS.find(m => m.openMetricName === 'vm_disk_throughput_write_bytes')
-    assert.ok(throughputWrite)
-    assert.equal(throughputWrite.type, 'gauge')
-    assert.ok(throughputWrite.transformValue)
-
-    const throughputTotal = VM_METRICS.find(m => m.openMetricName === 'vm_disk_throughput_total_bytes')
-    assert.ok(throughputTotal)
-    assert.equal(throughputTotal.type, 'gauge')
-    assert.ok(throughputTotal.transformValue)
-  })
-
-  it('should include VBD average latency metric with transformation', () => {
-    const latency = VM_METRICS.find(m => m.openMetricName === 'vm_disk_latency_seconds')
-    assert.ok(latency)
-    assert.equal(latency.type, 'gauge')
-    assert.ok(latency.transformValue)
-    // 500 µs / 1e6 = 0.0005 seconds
-    assert.equal(latency.transformValue!(500), 0.0005)
-
-    const regex = latency.test as RegExp
-    const match = regex.exec('vbd_xvda_latency')
-    assert.ok(match)
-    assert.deepEqual(latency.extractLabels!(match), { device: 'xvda' })
-  })
-
-  it('should not match vbd_xvda_latency with read_latency or write_latency regex', () => {
-    const readLatency = VM_METRICS.find(m => m.openMetricName === 'vm_disk_read_latency_seconds')
-    assert.ok(readLatency)
-    const readRegex = readLatency.test as RegExp
-    assert.equal(readRegex.exec('vbd_xvda_latency'), null)
-
-    const writeLatency = VM_METRICS.find(m => m.openMetricName === 'vm_disk_write_latency_seconds')
-    assert.ok(writeLatency)
-    const writeRegex = writeLatency.test as RegExp
-    assert.equal(writeRegex.exec('vbd_xvda_latency'), null)
-  })
-})
-
-describe('HOST_METRICS DCMI', () => {
-  it('should include DCMI power reading metric', () => {
-    const dcmi = HOST_METRICS.find(m => m.openMetricName === 'host_power_consumption_watts')
-    assert.ok(dcmi)
-    assert.equal(dcmi.type, 'gauge')
-    assert.equal(dcmi.test, 'DCMI-power-reading')
-    assert.equal(dcmi.transformValue, undefined)
-    assert.equal(dcmi.extractLabels, undefined)
-  })
-})
-
-describe('HOST_METRICS new metrics', () => {
-  it('should include hostload metric', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_load')
-    assert.ok(metric)
-    assert.equal(metric.type, 'gauge')
-    assert.equal(metric.test, 'hostload')
-    assert.equal(metric.transformValue, undefined)
-  })
-
-  it('should include memory_reclaimed metric with KiB to bytes transformation', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_memory_reclaimed_bytes')
-    assert.ok(metric)
-    assert.equal(metric.type, 'gauge')
-    assert.equal(metric.test, 'memory_reclaimed')
-    assert.ok(metric.transformValue)
-    // 512 KiB * 1024 = 524288 bytes
-    assert.equal(metric.transformValue!(512), 512 * 1024)
-  })
-
-  it('should include memory_reclaimed_max metric with KiB to bytes transformation', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_memory_reclaimed_max_bytes')
-    assert.ok(metric)
-    assert.equal(metric.type, 'gauge')
-    assert.equal(metric.test, 'memory_reclaimed_max')
-    assert.ok(metric.transformValue)
-    assert.equal(metric.transformValue!(1024), 1024 * 1024)
-  })
-
-  it('should include running_vcpus metric', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_running_vcpus')
-    assert.ok(metric)
-    assert.equal(metric.type, 'gauge')
-    assert.equal(metric.test, 'running_vcpus')
-    assert.equal(metric.transformValue, undefined)
-  })
-
-  it('should include pif_aggr_rx metric', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_network_aggregated_receive_bytes')
-    assert.ok(metric)
-    assert.equal(metric.type, 'gauge')
-    assert.equal(metric.test, 'pif_aggr_rx')
-    assert.equal(metric.transformValue, undefined)
-  })
-
-  it('should include pif_aggr_tx metric', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_network_aggregated_transmit_bytes')
-    assert.ok(metric)
-    assert.equal(metric.type, 'gauge')
-    assert.equal(metric.test, 'pif_aggr_tx')
-    assert.equal(metric.transformValue, undefined)
-  })
-
-  it('should match pif_aggr_rx before the generic PIF regex', () => {
-    const result = findMetricDefinition('pif_aggr_rx', 'host')
-    assert.ok(result)
-    assert.equal(result.definition.openMetricName, 'host_network_aggregated_receive_bytes')
-  })
-
-  it('should include iops_total per SR with label extraction', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_disk_iops_total')
-    assert.ok(metric)
-    assert.equal(metric.type, 'gauge')
-    assert.ok(metric.extractLabels)
-
-    const regex = metric.test as RegExp
-    const match = regex.exec('iops_total_abc12345')
-    assert.ok(match)
-    assert.deepEqual(metric.extractLabels!(match), { sr: 'abc12345' })
-  })
-
-  it('should include io_throughput_total per SR with MiB to bytes transformation', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_disk_throughput_total_bytes')
-    assert.ok(metric)
-    assert.equal(metric.type, 'gauge')
-    assert.ok(metric.transformValue)
-    assert.ok(metric.extractLabels)
-
-    // 2 MiB/s = 2 * 2^20 bytes/s
-    assert.equal(metric.transformValue!(2), 2 * 2 ** 20)
-
-    const regex = metric.test as RegExp
-    const match = regex.exec('io_throughput_total_def-456-789')
-    assert.ok(match)
-    assert.deepEqual(metric.extractLabels!(match), { sr: 'def-456-789' })
-  })
-
-  it('should include latency per SR with µs to seconds transformation', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_disk_latency_seconds')
-    assert.ok(metric)
-    assert.equal(metric.type, 'gauge')
-    assert.ok(metric.transformValue)
-    assert.ok(metric.extractLabels)
-
-    // 500 µs / 1e6 = 0.0005 seconds
-    assert.equal(metric.transformValue!(500), 0.0005)
-
-    const regex = metric.test as RegExp
-    const match = regex.exec('latency_abc-def-123')
-    assert.ok(match)
-    assert.deepEqual(metric.extractLabels!(match), { sr: 'abc-def-123' })
-  })
-
-  it('should not match read_latency or write_latency with latency_<sr> regex', () => {
-    const metric = HOST_METRICS.find(m => m.openMetricName === 'host_disk_latency_seconds')
-    assert.ok(metric)
-    const regex = metric.test as RegExp
-    // ^latency_ anchor prevents matching read_latency_ and write_latency_
-    assert.equal(regex.exec('read_latency_abc12345'), null)
-    assert.equal(regex.exec('write_latency_abc12345'), null)
-  })
-})
-
-// ============================================================================
-// findMetricDefinition Tests
-// ============================================================================
-
-describe('findMetricDefinition', () => {
-  it('should find host loadavg metric', () => {
-    const result = findMetricDefinition('loadavg', 'host')
-    assert.ok(result)
-    assert.equal(result.definition.openMetricName, 'host_load_average')
-  })
-
-  it('should find host CPU core metric and return matches', () => {
-    const result = findMetricDefinition('cpu3', 'host')
-    assert.ok(result)
-    assert.equal(result.definition.openMetricName, 'host_cpu_core_usage')
-    assert.ok(result.matches)
-    assert.equal(result.matches[1], '3')
-  })
-
-  it('should find VM cpu_usage metric', () => {
-    const result = findMetricDefinition('cpu_usage', 'vm')
-    assert.ok(result)
-    assert.equal(result.definition.openMetricName, 'vm_cpu_usage')
-  })
-
-  it('should find VM VIF metric', () => {
-    const result = findMetricDefinition('vif_2_tx', 'vm')
-    assert.ok(result)
-    assert.equal(result.definition.openMetricName, 'vm_network_transmit_bytes_total')
-    assert.ok(result.matches)
-    assert.equal(result.matches[1], '2')
-  })
-
-  it('should find VM VBD metric', () => {
-    const result = findMetricDefinition('vbd_xvdb_write', 'vm')
-    assert.ok(result)
-    assert.equal(result.definition.openMetricName, 'vm_disk_write_bytes_total')
-    assert.ok(result.matches)
-    assert.equal(result.matches[1], 'b')
-  })
-
-  it('should find VM VBD throughput metric', () => {
-    const result = findMetricDefinition('vbd_xvdb_io_throughput_read', 'vm')
-    assert.ok(result)
-    assert.equal(result.definition.openMetricName, 'vm_disk_throughput_read_bytes')
-    assert.ok(result.matches)
-    assert.equal(result.matches[1], 'b')
-  })
-
-  it('should find VM VBD average latency metric', () => {
-    const result = findMetricDefinition('vbd_xvda_latency', 'vm')
-    assert.ok(result)
-    assert.equal(result.definition.openMetricName, 'vm_disk_latency_seconds')
-  })
-
-  it('should find host DCMI power reading metric', () => {
-    const result = findMetricDefinition('DCMI-power-reading', 'host')
-    assert.ok(result)
-    assert.equal(result.definition.openMetricName, 'host_power_consumption_watts')
-    assert.equal(result.matches, null)
-  })
-
-  it('should return null for unknown metric', () => {
-    const result = findMetricDefinition('unknown_metric', 'host')
-    assert.equal(result, null)
-  })
-
-  it('should return null for SR object type', () => {
-    const result = findMetricDefinition('iops_read', 'sr')
-    assert.equal(result, null)
-  })
-})
-
-// ============================================================================
-// transformMetric Tests
-// ============================================================================
-
-describe('transformMetric', () => {
-  it('should transform host CPU metric', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'cpu_avg',
-        rawLegend: 'AVERAGE:host:host-uuid-123:cpu_avg',
-      },
-      value: 0.75,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456')
-
-    assert.ok(result)
-    assert.equal(result.name, 'xcp_host_cpu_average')
-    assert.equal(result.type, 'gauge')
-    assert.equal(result.value, 0.75)
-    assert.equal(result.timestamp, 1700000000)
-    assert.equal(result.labels.pool_id, 'pool-456')
-    assert.equal(result.labels.uuid, 'host-uuid-123')
-    assert.equal(result.labels.type, 'host')
-  })
-
-  it('should apply value transformation for memory', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'memory_free_kib',
-        rawLegend: 'AVERAGE:host:host-uuid-123:memory_free_kib',
-      },
-      value: 1024, // 1024 KiB
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456')
-
-    assert.ok(result)
-    assert.equal(result.name, 'xcp_host_memory_free_bytes')
-    assert.equal(result.value, 1024 * 1024) // 1 MiB in bytes
-  })
-
-  it('should extract labels from regex matches', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'cpu5',
-        rawLegend: 'AVERAGE:host:host-uuid-123:cpu5',
-      },
-      value: 0.25,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456')
-
-    assert.ok(result)
-    assert.equal(result.name, 'xcp_host_cpu_core_usage')
-    assert.equal(result.labels.core, '5')
-  })
-
-  it('should return null for null values', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'cpu_avg',
-        rawLegend: 'AVERAGE:host:host-uuid-123:cpu_avg',
-      },
-      value: null, // NaN was converted to null
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456')
-
-    assert.equal(result, null)
-  })
-
-  it('should return null for unknown metrics', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'unknown_metric',
-        rawLegend: 'AVERAGE:host:host-uuid-123:unknown_metric',
-      },
-      value: 42,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456')
-
-    assert.equal(result, null)
-  })
-
-  it('should transform VM VIF metric with extracted labels', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'vm-uuid-789',
-        metricName: 'vif_0_rx',
-        rawLegend: 'AVERAGE:vm:vm-uuid-789:vif_0_rx',
-      },
-      value: 1000000,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456')
-
-    assert.ok(result)
-    assert.equal(result.name, 'xcp_vm_network_receive_bytes_total')
-    assert.equal(result.type, 'counter')
-    assert.equal(result.labels.vif, '0')
-    assert.equal(result.labels.type, 'vm')
-  })
-})
-
-// ============================================================================
-// formatToOpenMetrics Tests
-// ============================================================================
-
-describe('formatToOpenMetrics', () => {
-  it('should format metrics to OpenMetrics text', () => {
-    const metrics: FormattedMetric[] = [
-      {
-        name: 'xcp_host_cpu_average',
-        help: 'Host average CPU usage ratio',
-        type: 'gauge',
-        labels: { pool_id: 'pool-1', uuid: 'host-1', type: 'host' },
-        value: 0.5,
-        timestamp: 1700000000,
-      },
-    ]
-
-    const result = formatToOpenMetrics(metrics)
-
-    assert.ok(result.includes('# HELP xcp_host_cpu_average Host average CPU usage ratio'))
-    assert.ok(result.includes('# TYPE xcp_host_cpu_average gauge'))
-    assert.ok(result.includes('xcp_host_cpu_average{pool_id="pool-1",uuid="host-1",type="host"} 0.5 1700000000'))
-  })
-
-  it('should group metrics by name', () => {
-    const metrics: FormattedMetric[] = [
-      {
-        name: 'xcp_host_cpu_core_usage',
-        help: 'Host CPU core usage ratio',
-        type: 'gauge',
-        labels: { pool_id: 'pool-1', uuid: 'host-1', type: 'host', core: '0' },
-        value: 0.3,
-        timestamp: 1700000000,
-      },
-      {
-        name: 'xcp_host_cpu_core_usage',
-        help: 'Host CPU core usage ratio',
-        type: 'gauge',
-        labels: { pool_id: 'pool-1', uuid: 'host-1', type: 'host', core: '1' },
-        value: 0.4,
-        timestamp: 1700000000,
-      },
-    ]
-
-    const result = formatToOpenMetrics(metrics)
-
-    // Should only have one HELP and TYPE declaration
-    const helpCount = (result.match(/# HELP xcp_host_cpu_core_usage/g) || []).length
-    const typeCount = (result.match(/# TYPE xcp_host_cpu_core_usage/g) || []).length
-
-    assert.equal(helpCount, 1)
-    assert.equal(typeCount, 1)
-
-    // Should have two value lines
-    assert.ok(result.includes('core="0"'))
-    assert.ok(result.includes('core="1"'))
-  })
-
-  it('should return empty string for empty metrics', () => {
-    const result = formatToOpenMetrics([])
-    assert.equal(result, '')
-  })
-
-  it('should escape special characters in label values', () => {
-    const metrics: FormattedMetric[] = [
-      {
-        name: 'xcp_test_metric',
-        help: 'Test metric',
-        type: 'gauge',
-        labels: { name: 'value with "quotes"' },
-        value: 1,
-        timestamp: 1700000000,
-      },
-    ]
-
-    const result = formatToOpenMetrics(metrics)
-
-    assert.ok(result.includes('name="value with \\"quotes\\""'))
-  })
-})
-
-// ============================================================================
-// formatAllPoolsToOpenMetrics Tests
-// ============================================================================
-
-describe('formatAllPoolsToOpenMetrics', () => {
-  it('should format RRD data from multiple pools', () => {
-    const rrdDataList: ParsedRrdData[] = [
-      {
-        poolId: 'pool-1',
-        timestamp: 1700000000,
-        metrics: [
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'host',
-              uuid: 'host-1',
-              metricName: 'cpu_avg',
-              rawLegend: 'AVERAGE:host:host-1:cpu_avg',
-            },
-            value: 0.5,
-            timestamp: 1700000000,
-          },
-        ],
-      },
-      {
-        poolId: 'pool-2',
-        timestamp: 1700000000,
-        metrics: [
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'host',
-              uuid: 'host-2',
-              metricName: 'cpu_avg',
-              rawLegend: 'AVERAGE:host:host-2:cpu_avg',
-            },
-            value: 0.7,
-            timestamp: 1700000000,
-          },
-        ],
-      },
-    ]
-
-    const result = formatAllPoolsToOpenMetrics(rrdDataList)
-
-    assert.ok(result.includes('pool_id="pool-1"'))
-    assert.ok(result.includes('pool_id="pool-2"'))
-    assert.ok(result.includes('# EOF'))
-  })
-
-  it('should return empty string for empty data', () => {
-    const result = formatAllPoolsToOpenMetrics([])
-
-    assert.equal(result, '')
-  })
-
-  it('should skip metrics with null values', () => {
-    const rrdDataList: ParsedRrdData[] = [
-      {
-        poolId: 'pool-1',
-        timestamp: 1700000000,
-        metrics: [
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'host',
-              uuid: 'host-1',
-              metricName: 'cpu_avg',
-              rawLegend: 'AVERAGE:host:host-1:cpu_avg',
-            },
-            value: null, // NaN converted to null
-            timestamp: 1700000000,
-          },
-        ],
-      },
-    ]
-
-    const result = formatAllPoolsToOpenMetrics(rrdDataList)
-
-    // Should return empty string since the metric has null value
-    assert.equal(result, '')
-  })
-
-  it('should sort metrics by name for consistent output', () => {
-    const rrdDataList: ParsedRrdData[] = [
-      {
-        poolId: 'pool-1',
-        timestamp: 1700000000,
-        metrics: [
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'host',
-              uuid: 'host-1',
-              metricName: 'memory_free_kib',
-              rawLegend: 'AVERAGE:host:host-1:memory_free_kib',
-            },
-            value: 1024,
-            timestamp: 1700000000,
-          },
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'host',
-              uuid: 'host-1',
-              metricName: 'cpu_avg',
-              rawLegend: 'AVERAGE:host:host-1:cpu_avg',
-            },
-            value: 0.5,
-            timestamp: 1700000000,
-          },
-        ],
-      },
-    ]
-
-    const result = formatAllPoolsToOpenMetrics(rrdDataList)
-    const lines = result.split('\n')
-
-    // cpu_average should come before memory_free_bytes alphabetically
-    const cpuIndex = lines.findIndex(l => l.includes('xcp_host_cpu_average'))
-    const memIndex = lines.findIndex(l => l.includes('xcp_host_memory_free_bytes'))
-
-    assert.ok(cpuIndex < memIndex, 'Metrics should be sorted alphabetically')
-  })
-})
-
-// ============================================================================
-// Label Context Tests
-// ============================================================================
-
-describe('transformMetric with labelContext', () => {
-  const createLabelContext = (): LabelContext => ({
-    hosts: [
-      {
-        hostId: 'host-uuid-123',
-        hostAddress: '192.168.1.1',
-        hostLabel: 'Host 1',
-        poolId: 'pool-456',
-        poolLabel: 'Production Pool',
-        sessionId: 'session-123',
-        protocol: 'https:',
-      },
-    ],
-    labels: {
-      vms: {
-        'vm-uuid-789': {
-          name_label: 'Web Server',
-          is_control_domain: false,
-          vbdDeviceToVdiName: { xvda: 'System Disk', xvdb: 'Data Disk' },
-          vbdDeviceToVdiUuid: { xvda: 'vdi-uuid-system', xvdb: 'vdi-uuid-data' },
-          vifIndexToNetworkName: { '0': 'Pool-wide network', '1': 'Storage network' },
-          startTime: null,
-          power_state: 'Running',
-          pool_id: 'pool-456',
-          pool_name: 'Production Pool',
-        },
-      },
-      hosts: {
-        'host-uuid-123': {
-          name_label: 'Host 1',
-          pifDeviceToNetworkName: { eth0: 'Pool-wide network', eth1: 'Storage network' },
-          startTime: null,
-        },
-      },
-      srs: {
-        'sr-uuid-full-1234567890': {
-          name_label: 'Local Storage',
-          SR_type: 'lvm',
-        },
-      },
-      srTruncatedToUuid: {
-        '1234567890': 'sr-uuid-full-1234567890',
-      },
-      vdiUuidToSrUuid: {
-        'vdi-uuid-system': 'sr-uuid-full-1234567890',
-        'vdi-uuid-data': 'sr-uuid-full-1234567890',
-      },
-    },
-  })
-
-  it('should add pool_name label from host credentials', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'cpu_avg',
-        rawLegend: 'AVERAGE:host:host-uuid-123:cpu_avg',
-      },
-      value: 0.75,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.pool_name, 'Production Pool')
-  })
-
-  it('should add host_name label for host metrics', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'cpu_avg',
-        rawLegend: 'AVERAGE:host:host-uuid-123:cpu_avg',
-      },
-      value: 0.75,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.host_name, 'Host 1')
-  })
-
-  it('should add vm_name label for VM metrics', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'vm-uuid-789',
-        metricName: 'cpu_usage',
-        rawLegend: 'AVERAGE:vm:vm-uuid-789:cpu_usage',
-      },
-      value: 0.5,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.vm_name, 'Web Server')
-  })
-
-  it('should add vdi_name label for VBD metrics', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'vm-uuid-789',
-        metricName: 'vbd_xvda_read',
-        rawLegend: 'AVERAGE:vm:vm-uuid-789:vbd_xvda_read',
-      },
-      value: 1000000,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.device, 'xvda')
-    assert.equal(result.labels.vdi_name, 'System Disk')
-  })
-
-  it('should add sr_name label for VM disk metrics via VDI-SR mapping', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'vm-uuid-789',
-        metricName: 'vbd_xvda_iops_read',
-        rawLegend: 'AVERAGE:vm:vm-uuid-789:vbd_xvda_iops_read',
-      },
-      value: 150,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.device, 'xvda')
-    assert.equal(result.labels.vdi_name, 'System Disk')
-    assert.equal(result.labels.sr_name, 'Local Storage')
-  })
-
-  it('should add network_name label for VIF metrics', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'vm-uuid-789',
-        metricName: 'vif_0_rx',
-        rawLegend: 'AVERAGE:vm:vm-uuid-789:vif_0_rx',
-      },
-      value: 500000,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.vif, '0')
-    assert.equal(result.labels.network_name, 'Pool-wide network')
-  })
-
-  it('should add network_name label for PIF metrics', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'pif_eth0_rx',
-        rawLegend: 'AVERAGE:host:host-uuid-123:pif_eth0_rx',
-      },
-      value: 1000000,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.interface, 'eth0')
-    assert.equal(result.labels.network_name, 'Pool-wide network')
-  })
-
-  it('should add sr_name label for SR metrics', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'iops_read_1234567890',
-        rawLegend: 'AVERAGE:host:host-uuid-123:iops_read_1234567890',
-      },
-      value: 100,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.sr, '1234567890')
-    assert.equal(result.labels.sr_name, 'Local Storage')
-  })
-
-  it('should add sr_type label for host SR-tagged metrics (iops/throughput/latency)', () => {
-    const ctx = createLabelContext()
-    ctx.labels.srs['sr-uuid-full-1234567890']!.SR_type = 'linstor'
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'iops_read_1234567890',
-        rawLegend: 'AVERAGE:host:host-uuid-123:iops_read_1234567890',
-      },
-      value: 100,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', ctx)
-
-    assert.ok(result)
-    assert.equal(result.labels.sr_type, 'linstor')
-  })
-
-  it('should add sr_type label for VBD VM disk metrics via VDI→SR mapping', () => {
-    const ctx = createLabelContext()
-    ctx.labels.srs['sr-uuid-full-1234567890']!.SR_type = 'linstor'
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'vm-uuid-789',
-        metricName: 'vbd_xvda_iops_read',
-        rawLegend: 'AVERAGE:vm:vm-uuid-789:vbd_xvda_iops_read',
-      },
-      value: 200,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', ctx)
-
-    assert.ok(result)
-    assert.equal(result.labels.sr_type, 'linstor')
-  })
-
-  it('should omit sr_type label when SR_type is empty', () => {
-    const ctx = createLabelContext()
-    ctx.labels.srs['sr-uuid-full-1234567890']!.SR_type = ''
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'iops_read_1234567890',
-        rawLegend: 'AVERAGE:host:host-uuid-123:iops_read_1234567890',
-      },
-      value: 100,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', ctx)
-
-    assert.ok(result)
-    assert.equal(result.labels.sr_type, undefined)
-  })
-
-  it('should resolve sr_name and sr_type from a UUID prefix (XCP-ng RRD encoding)', () => {
-    // XCP-ng encodes the SR in RRD legends as the first 8 chars of the UUID.
-    // The label context must therefore index that prefix.
-    const fullUuid = 'c787b75c-3e0d-70fa-d0c3-cbfd382d7e33'
-    const prefix = fullUuid.slice(0, 8)
-    const ctx = createLabelContext()
-    ctx.labels.srs[fullUuid] = { name_label: 'XOSTOR NVME', SR_type: 'linstor' }
-    ctx.labels.srTruncatedToUuid[prefix] = fullUuid
-
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: `iops_read_${prefix}`,
-        rawLegend: `AVERAGE:host:host-uuid-123:iops_read_${prefix}`,
-      },
-      value: 42,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', ctx)
-
-    assert.ok(result)
-    assert.equal(result.labels.sr, prefix)
-    assert.equal(result.labels.sr_name, 'XOSTOR NVME')
-    assert.equal(result.labels.sr_type, 'linstor')
-  })
-
-  it('should handle missing label context gracefully', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'host',
-        uuid: 'host-uuid-123',
-        metricName: 'cpu_avg',
-        rawLegend: 'AVERAGE:host:host-uuid-123:cpu_avg',
-      },
-      value: 0.75,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456')
-
-    assert.ok(result)
-    assert.equal(result.labels.pool_name, undefined)
-    assert.equal(result.labels.host_name, undefined)
-  })
-
-  it('should add is_control_domain="false" label for regular VMs', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'vm-uuid-789',
-        metricName: 'cpu_usage',
-        rawLegend: 'AVERAGE:vm:vm-uuid-789:cpu_usage',
-      },
-      value: 0.5,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.is_control_domain, 'false')
-  })
-
-  it('should add is_control_domain="true" label for dom0 VMs', () => {
-    const context = createLabelContext()
-    context.labels.vms['dom0-uuid'] = {
-      name_label: 'Control domain on host: Host 1',
-      is_control_domain: true,
-      vbdDeviceToVdiName: {},
-      vbdDeviceToVdiUuid: {},
-      vifIndexToNetworkName: {},
-      startTime: null,
-      power_state: 'Running',
-      pool_id: 'pool-456',
-      pool_name: 'Production Pool',
-    }
-
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'dom0-uuid',
-        metricName: 'cpu_usage',
-        rawLegend: 'AVERAGE:vm:dom0-uuid:cpu_usage',
-      },
-      value: 0.3,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', context)
-
-    assert.ok(result)
-    assert.equal(result.labels.is_control_domain, 'true')
-    assert.equal(result.labels.vm_name, 'Control domain on host: Host 1')
-  })
-
-  it('should handle missing VM in label context gracefully', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'unknown-vm-uuid',
-        metricName: 'cpu_usage',
-        rawLegend: 'AVERAGE:vm:unknown-vm-uuid:cpu_usage',
-      },
-      value: 0.5,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.vm_name, undefined)
-    assert.equal(result.labels.is_control_domain, undefined)
-  })
-
-  it('should handle missing VDI mapping gracefully', () => {
-    const metric: ParsedMetric = {
-      legend: {
-        cf: 'AVERAGE',
-        objectType: 'vm',
-        uuid: 'vm-uuid-789',
-        metricName: 'vbd_xvdc_read',
-        rawLegend: 'AVERAGE:vm:vm-uuid-789:vbd_xvdc_read',
-      },
-      value: 1000000,
-      timestamp: 1700000000,
-    }
-
-    const result = transformMetric(metric, 'pool-456', createLabelContext())
-
-    assert.ok(result)
-    assert.equal(result.labels.device, 'xvdc')
-    assert.equal(result.labels.vdi_name, undefined)
-  })
-})
-
-describe('formatAllPoolsToOpenMetrics with labelContext', () => {
-  const createLabelContext = (): LabelContext => ({
-    hosts: [
-      {
-        hostId: 'host-1',
-        hostAddress: '192.168.1.1',
-        hostLabel: 'Host 1',
-        poolId: 'pool-1',
-        poolLabel: 'Production',
-        sessionId: 'session-1',
-        protocol: 'https:',
-      },
-    ],
-    labels: {
-      vms: {
-        'vm-1': {
-          name_label: 'Web Server',
-          is_control_domain: false,
-          vbdDeviceToVdiName: { xvda: 'System Disk' },
-          vbdDeviceToVdiUuid: { xvda: 'vdi-1' },
-          vifIndexToNetworkName: { '0': 'Management' },
-          startTime: null,
-          power_state: 'Running',
-          pool_id: 'pool-1',
-          pool_name: 'Production',
-        },
-      },
-      hosts: {
-        'host-1': {
-          name_label: 'Host 1',
-          pifDeviceToNetworkName: { eth0: 'Management' },
-          startTime: null,
-        },
-      },
-      srs: {},
-      srTruncatedToUuid: {},
-      vdiUuidToSrUuid: {},
-    },
-  })
-
-  it('should include human-readable labels in formatted output', () => {
-    const rrdDataList: ParsedRrdData[] = [
-      {
-        poolId: 'pool-1',
-        timestamp: 1700000000,
-        metrics: [
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'host',
-              uuid: 'host-1',
-              metricName: 'cpu_avg',
-              rawLegend: 'AVERAGE:host:host-1:cpu_avg',
-            },
-            value: 0.5,
-            timestamp: 1700000000,
-          },
-        ],
-      },
-    ]
-
-    const result = formatAllPoolsToOpenMetrics(rrdDataList, createLabelContext())
-
-    assert.ok(result.includes('pool_name="Production"'))
-    assert.ok(result.includes('host_name="Host 1"'))
-  })
-
-  it('should include VM labels in VM metrics', () => {
-    const rrdDataList: ParsedRrdData[] = [
-      {
-        poolId: 'pool-1',
-        timestamp: 1700000000,
-        metrics: [
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-1',
-              metricName: 'cpu_usage',
-              rawLegend: 'AVERAGE:vm:vm-1:cpu_usage',
-            },
-            value: 0.75,
-            timestamp: 1700000000,
-          },
-        ],
-      },
-    ]
-
-    const result = formatAllPoolsToOpenMetrics(rrdDataList, createLabelContext())
-
-    assert.ok(result.includes('vm_name="Web Server"'))
-  })
-})
-
-// ============================================================================
-// CPU Usage Fallback Tests
-// ============================================================================
-
-describe('CPU usage fallback', () => {
-  it('should not generate fallback when vm_cpu_usage is present', () => {
-    const rrdDataList: ParsedRrdData[] = [
-      {
-        poolId: 'pool-1',
-        timestamp: 1700000000,
-        metrics: [
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-1',
-              metricName: 'cpu_usage',
-              rawLegend: 'AVERAGE:vm:vm-1:cpu_usage',
-            },
-            value: 0.5,
-            timestamp: 1700000000,
-          },
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-1',
-              metricName: 'cpu0',
-              rawLegend: 'AVERAGE:vm:vm-1:cpu0',
-            },
-            value: 0.6,
-            timestamp: 1700000000,
-          },
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-1',
-              metricName: 'cpu1',
-              rawLegend: 'AVERAGE:vm:vm-1:cpu1',
-            },
-            value: 0.4,
-            timestamp: 1700000000,
-          },
-        ],
-      },
-    ]
-
-    const result = formatAllPoolsToOpenMetrics(rrdDataList)
-    const lines = result.split('\n')
-
-    // Should have exactly one vm_cpu_usage metric (the native one, not synthetic)
-    const cpuUsageLines = lines.filter(l => l.includes('xcp_vm_cpu_usage{'))
-    assert.equal(cpuUsageLines.length, 1)
-    assert.ok(cpuUsageLines[0]?.includes('0.5'))
-  })
-
-  it('should generate fallback vm_cpu_usage from per-core metrics', () => {
-    const rrdDataList: ParsedRrdData[] = [
-      {
-        poolId: 'pool-1',
-        timestamp: 1700000000,
-        metrics: [
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-1',
-              metricName: 'cpu0',
-              rawLegend: 'AVERAGE:vm:vm-1:cpu0',
-            },
-            value: 0.6,
-            timestamp: 1700000000,
-          },
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-1',
-              metricName: 'cpu1',
-              rawLegend: 'AVERAGE:vm:vm-1:cpu1',
-            },
-            value: 0.4,
-            timestamp: 1700000000,
-          },
-        ],
-      },
-    ]
-
-    const result = formatAllPoolsToOpenMetrics(rrdDataList)
-
-    // Should have vm_cpu_usage metric generated from average of cores
-    assert.ok(result.includes('xcp_vm_cpu_usage'))
-
-    const lines = result.split('\n')
-    const cpuUsageLine = lines.find(l => l.includes('xcp_vm_cpu_usage{'))
-    assert.ok(cpuUsageLine)
-    // Average of 0.6 and 0.4 = 0.5
-    assert.ok(cpuUsageLine.includes('0.5'))
-  })
-
-  it('should generate fallback for multiple VMs', () => {
-    const rrdDataList: ParsedRrdData[] = [
-      {
-        poolId: 'pool-1',
-        timestamp: 1700000000,
-        metrics: [
-          // VM 1: only per-core metrics
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-1',
-              metricName: 'cpu0',
-              rawLegend: 'AVERAGE:vm:vm-1:cpu0',
-            },
-            value: 0.8,
-            timestamp: 1700000000,
-          },
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-1',
-              metricName: 'cpu1',
-              rawLegend: 'AVERAGE:vm:vm-1:cpu1',
-            },
-            value: 0.2,
-            timestamp: 1700000000,
-          },
-          // VM 2: has native cpu_usage (should not get fallback)
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-2',
-              metricName: 'cpu_usage',
-              rawLegend: 'AVERAGE:vm:vm-2:cpu_usage',
-            },
-            value: 0.3,
-            timestamp: 1700000000,
-          },
-          // VM 3: only per-core metrics
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-3',
-              metricName: 'cpu0',
-              rawLegend: 'AVERAGE:vm:vm-3:cpu0',
-            },
-            value: 0.9,
-            timestamp: 1700000000,
-          },
-        ],
-      },
-    ]
-
-    const result = formatAllPoolsToOpenMetrics(rrdDataList)
-    const lines = result.split('\n')
-
-    // VM 1 should have synthetic cpu_usage (average of 0.8 and 0.2 = 0.5)
-    const vm1CpuUsage = lines.find(l => l.includes('xcp_vm_cpu_usage') && l.includes('uuid="vm-1"'))
-    assert.ok(vm1CpuUsage)
-    assert.ok(vm1CpuUsage.includes('0.5'))
-
-    // VM 2 should have its native cpu_usage (0.3)
-    const vm2CpuUsage = lines.find(l => l.includes('xcp_vm_cpu_usage') && l.includes('uuid="vm-2"'))
-    assert.ok(vm2CpuUsage)
-    assert.ok(vm2CpuUsage.includes('0.3'))
-
-    // VM 3 should have synthetic cpu_usage (0.9 single core)
-    const vm3CpuUsage = lines.find(l => l.includes('xcp_vm_cpu_usage') && l.includes('uuid="vm-3"'))
-    assert.ok(vm3CpuUsage)
-    assert.ok(vm3CpuUsage.includes('0.9'))
-  })
-
-  it('should preserve labels in fallback metric', () => {
-    const labelContext: LabelContext = {
-      hosts: [
-        {
-          hostId: 'host-1',
-          hostAddress: '192.168.1.1',
-          hostLabel: 'Host 1',
-          poolId: 'pool-1',
-          poolLabel: 'Production',
-          sessionId: 'session-1',
-          protocol: 'https:',
-        },
-      ],
-      labels: {
-        vms: {
-          'vm-1': {
-            name_label: 'XCP-ng 8.2 VM',
-            is_control_domain: false,
-            vbdDeviceToVdiName: {},
-            vbdDeviceToVdiUuid: {},
-            vifIndexToNetworkName: {},
-            startTime: null,
-            power_state: 'Running',
-            pool_id: 'pool-1',
-            pool_name: 'Production',
-          },
-        },
-        hosts: {},
-        srs: {},
-        srTruncatedToUuid: {},
-        vdiUuidToSrUuid: {},
-      },
-    }
-
-    const rrdDataList: ParsedRrdData[] = [
-      {
-        poolId: 'pool-1',
-        timestamp: 1700000000,
-        metrics: [
-          {
-            legend: {
-              cf: 'AVERAGE',
-              objectType: 'vm',
-              uuid: 'vm-1',
-              metricName: 'cpu0',
-              rawLegend: 'AVERAGE:vm:vm-1:cpu0',
-            },
-            value: 0.75,
-            timestamp: 1700000000,
-          },
-        ],
-      },
-    ]
-
-    const result = formatAllPoolsToOpenMetrics(rrdDataList, labelContext)
-
-    // Fallback metric should have all the labels (pool_id, uuid, type, vm_name, pool_name)
-    // but NOT the core label
-    assert.ok(result.includes('xcp_vm_cpu_usage'))
-    assert.ok(result.includes('pool_id="pool-1"'))
-    assert.ok(result.includes('uuid="vm-1"'))
-    assert.ok(result.includes('type="vm"'))
-    assert.ok(result.includes('vm_name="XCP-ng 8.2 VM"'))
-    assert.ok(result.includes('pool_name="Production"'))
-
-    // Should NOT have core label in the fallback metric
-    const lines = result.split('\n')
-    const cpuUsageLine = lines.find(l => l.includes('xcp_vm_cpu_usage{'))
-    assert.ok(cpuUsageLine)
-    assert.ok(!cpuUsageLine.includes('core='))
-  })
-})
-
-// ============================================================================
-// formatHostStatusMetrics Tests
-// ============================================================================
-
-describe('formatHostStatusMetrics', () => {
-  it('should return empty array for empty input', () => {
-    const result = formatHostStatusMetrics([])
-    assert.deepEqual(result, [])
-  })
-
-  it('should create one metric per host with value 1', () => {
-    const hosts: HostStatusItem[] = [
-      {
-        uuid: 'host-1',
-        name_label: 'Host 1',
-        power_state: 'Running',
-        enabled: true,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-      {
-        uuid: 'host-2',
-        name_label: 'Host 2',
-        power_state: 'Halted',
-        enabled: true,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-    ]
-
-    const result = formatHostStatusMetrics(hosts)
-
-    assert.equal(result.length, 2)
-    assert.equal(result[0]!.value, 1)
-    assert.equal(result[1]!.value, 1)
-  })
-
-  it('should set correct metric name and type', () => {
-    const hosts: HostStatusItem[] = [
-      {
-        uuid: 'host-1',
-        name_label: 'Host 1',
-        power_state: 'Running',
-        enabled: true,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-    ]
-
-    const result = formatHostStatusMetrics(hosts)
-
-    assert.equal(result[0]!.name, 'xcp_host_status')
-    assert.equal(result[0]!.type, 'gauge')
-  })
-
-  it('should include all expected labels', () => {
-    const hosts: HostStatusItem[] = [
-      {
-        uuid: 'host-1',
-        name_label: 'Host 1',
-        power_state: 'Running',
-        enabled: true,
-        pool_id: 'pool-1',
-        pool_name: 'Production',
-      },
-    ]
-
-    const result = formatHostStatusMetrics(hosts)
-    const labels = result[0]!.labels
-
-    assert.equal(labels.pool_id, 'pool-1')
-    assert.equal(labels.pool_name, 'Production')
-    assert.equal(labels.uuid, 'host-1')
-    assert.equal(labels.host_name, 'Host 1')
-    assert.equal(labels.power_state, 'Running')
-    assert.equal(labels.enabled, 'true')
-  })
-
-  it('should omit pool_name when empty', () => {
-    const hosts: HostStatusItem[] = [
-      { uuid: 'host-1', name_label: 'Host 1', power_state: 'Unknown', enabled: true, pool_id: 'pool-1', pool_name: '' },
-    ]
-
-    const result = formatHostStatusMetrics(hosts)
-
-    assert.equal(result[0]!.labels.pool_name, undefined)
-  })
-
-  it('should handle all power_state values and enabled flag', () => {
-    const hosts: HostStatusItem[] = [
-      {
-        uuid: 'host-0',
-        name_label: 'Host 0',
-        power_state: 'Running',
-        enabled: true,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-      {
-        uuid: 'host-1',
-        name_label: 'Host 1',
-        power_state: 'Running',
-        enabled: false,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-      {
-        uuid: 'host-2',
-        name_label: 'Host 2',
-        power_state: 'Halted',
-        enabled: true,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-      {
-        uuid: 'host-3',
-        name_label: 'Host 3',
-        power_state: 'Unknown',
-        enabled: true,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-    ]
-
-    const result = formatHostStatusMetrics(hosts)
-
-    assert.equal(result.length, 4)
-    assert.equal(result[0]!.labels.power_state, 'Running')
-    assert.equal(result[0]!.labels.enabled, 'true')
-    assert.equal(result[1]!.labels.power_state, 'Running')
-    assert.equal(result[1]!.labels.enabled, 'false')
-    assert.equal(result[2]!.labels.power_state, 'Halted')
-    assert.equal(result[3]!.labels.power_state, 'Unknown')
-  })
-
-  it('should produce valid OpenMetrics output', () => {
-    const hosts: HostStatusItem[] = [
-      {
-        uuid: 'host-1',
-        name_label: 'Host 1',
-        power_state: 'Running',
-        enabled: true,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-      {
-        uuid: 'host-2',
-        name_label: 'Host 2',
-        power_state: 'Halted',
-        enabled: false,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-    ]
-
-    const metrics = formatHostStatusMetrics(hosts)
-    const output = formatToOpenMetrics(metrics)
-
-    assert.ok(output.includes('# HELP xcp_host_status Host status (1 = current state)'))
-    assert.ok(output.includes('# TYPE xcp_host_status gauge'))
-    assert.ok(output.includes('power_state="Running"'))
-    assert.ok(output.includes('enabled="false"'))
-
-    // HELP and TYPE should appear only once
-    const helpCount = (output.match(/# HELP xcp_host_status/g) || []).length
-    assert.equal(helpCount, 1)
-  })
-
-  it('should escape special characters in host names', () => {
-    const hosts: HostStatusItem[] = [
-      {
-        uuid: 'host-1',
-        name_label: 'Host "with quotes"',
-        power_state: 'Running',
-        enabled: true,
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-    ]
-
-    const metrics = formatHostStatusMetrics(hosts)
-    const output = formatToOpenMetrics(metrics)
-
-    assert.ok(output.includes('host_name="Host \\"with quotes\\""'))
-  })
-})
-
-// ============================================================================
-// Host Uptime Metrics Tests
-// ============================================================================
-
-describe('formatHostUptimeMetrics', () => {
-  const createLabelContextWithUptime = (startTime: number | null): LabelContext => ({
-    hosts: [
-      {
-        hostId: 'host-uuid-123',
-        hostAddress: '192.168.1.1',
-        hostLabel: 'Host 1',
-        poolId: 'pool-456',
-        poolLabel: 'Production Pool',
-        sessionId: 'session-123',
-        protocol: 'https:',
-      },
-    ],
-    labels: {
-      vms: {},
-      hosts: {
-        'host-uuid-123': {
-          name_label: 'Host 1',
-          pifDeviceToNetworkName: {},
-          startTime,
-        },
-      },
-      srs: {},
-      srTruncatedToUuid: {},
-      vdiUuidToSrUuid: {},
-    },
-  })
-
-  it('should generate uptime metric for host with valid startTime', () => {
-    const now = Math.floor(Date.now() / 1000)
-    const bootTime = now - 3600 // 1 hour ago
-    const labelContext = createLabelContextWithUptime(bootTime)
-
-    const metrics = formatHostUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 1)
-    const metric = metrics[0]!
-    assert.equal(metric.name, 'xcp_host_uptime_seconds')
-    assert.equal(metric.type, 'gauge')
-    assert.equal(metric.help, 'Host uptime in seconds since boot')
-    assert.equal(metric.labels.pool_id, 'pool-456')
-    assert.equal(metric.labels.pool_name, 'Production Pool')
-    assert.equal(metric.labels.uuid, 'host-uuid-123')
-    assert.equal(metric.labels.host_name, 'Host 1')
-    // Value should be approximately 3600 (1 hour)
-    assert.ok(metric.value >= 3599 && metric.value <= 3601)
-  })
-
-  it('should skip host with null startTime', () => {
-    const labelContext = createLabelContextWithUptime(null)
-
-    const metrics = formatHostUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 0)
-  })
-
-  it('should skip host not found in labels', () => {
-    const labelContext: LabelContext = {
-      hosts: [
-        {
-          hostId: 'unknown-host',
-          hostAddress: '192.168.1.1',
-          hostLabel: 'Unknown Host',
-          poolId: 'pool-456',
-          poolLabel: 'Production',
-          sessionId: 'session-123',
-          protocol: 'https:',
-        },
-      ],
-      labels: {
-        vms: {},
-        hosts: {},
-        srs: {},
-        srTruncatedToUuid: {},
-        vdiUuidToSrUuid: {},
-      },
-    }
-
-    const metrics = formatHostUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 0)
-  })
-
-  it('should generate metrics for multiple hosts', () => {
-    const now = Math.floor(Date.now() / 1000)
-    const labelContext: LabelContext = {
-      hosts: [
-        {
-          hostId: 'host-1',
-          hostAddress: '192.168.1.1',
-          hostLabel: 'Host 1',
-          poolId: 'pool-1',
-          poolLabel: 'Pool A',
-          sessionId: 'session-1',
-          protocol: 'https:',
-        },
-        {
-          hostId: 'host-2',
-          hostAddress: '192.168.1.2',
-          hostLabel: 'Host 2',
-          poolId: 'pool-1',
-          poolLabel: 'Pool A',
-          sessionId: 'session-1',
-          protocol: 'https:',
-        },
-      ],
-      labels: {
-        vms: {},
-        hosts: {
-          'host-1': {
-            name_label: 'Host 1',
-            pifDeviceToNetworkName: {},
-            startTime: now - 7200, // 2 hours
-          },
-          'host-2': {
-            name_label: 'Host 2',
-            pifDeviceToNetworkName: {},
-            startTime: now - 1800, // 30 minutes
-          },
-        },
-        srs: {},
-        srTruncatedToUuid: {},
-        vdiUuidToSrUuid: {},
-      },
-    }
-
-    const metrics = formatHostUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 2)
-    const host1Metric = metrics.find(m => m.labels.uuid === 'host-1')
-    const host2Metric = metrics.find(m => m.labels.uuid === 'host-2')
-    assert.ok(host1Metric)
-    assert.ok(host2Metric)
-    assert.ok(host1Metric.value >= 7199 && host1Metric.value <= 7201)
-    assert.ok(host2Metric.value >= 1799 && host2Metric.value <= 1801)
-  })
-
-  it('should omit pool_name label when empty', () => {
-    const now = Math.floor(Date.now() / 1000)
-    const labelContext: LabelContext = {
-      hosts: [
-        {
-          hostId: 'host-1',
-          hostAddress: '192.168.1.1',
-          hostLabel: 'Host 1',
-          poolId: 'pool-1',
-          poolLabel: '', // Empty pool label
-          sessionId: 'session-1',
-          protocol: 'https:',
-        },
-      ],
-      labels: {
-        vms: {},
-        hosts: {
-          'host-1': {
-            name_label: 'Host 1',
-            pifDeviceToNetworkName: {},
-            startTime: now - 3600,
-          },
-        },
-        srs: {},
-        srTruncatedToUuid: {},
-        vdiUuidToSrUuid: {},
-      },
-    }
-
-    const metrics = formatHostUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 1)
-    assert.equal(metrics[0]!.labels.pool_name, undefined)
-  })
-
-  it('should omit host_name label when empty', () => {
-    const now = Math.floor(Date.now() / 1000)
-    const labelContext: LabelContext = {
-      hosts: [
-        {
-          hostId: 'host-1',
-          hostAddress: '192.168.1.1',
-          hostLabel: 'Host 1',
-          poolId: 'pool-1',
-          poolLabel: 'Pool A',
-          sessionId: 'session-1',
-          protocol: 'https:',
-        },
-      ],
-      labels: {
-        vms: {},
-        hosts: {
-          'host-1': {
-            name_label: '', // Empty host name
-            pifDeviceToNetworkName: {},
-            startTime: now - 3600,
-          },
-        },
-        srs: {},
-        srTruncatedToUuid: {},
-        vdiUuidToSrUuid: {},
-      },
-    }
-
-    const metrics = formatHostUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 1)
-    assert.equal(metrics[0]!.labels.host_name, undefined)
-  })
-})
-
-// ============================================================================
-// SR Metrics Tests
-// ============================================================================
-
-describe('formatSrMetrics', () => {
-  const createSrDataItem = (overrides: Partial<SrDataItem> = {}): SrDataItem => ({
-    uuid: 'sr-uuid-456',
-    name_label: 'Local Storage',
-    size: 1_000_000_000_000,
-    physical_usage: 250_000_000_000,
-    usage: 500_000_000_000,
-    pool_id: 'pool-789',
-    pool_name: 'Production Pool',
-    sr_type: 'lvm',
-    ...overrides,
-  })
-
-  it('should emit 3 metrics per SR with sr_type label when present', () => {
-    const metrics = formatSrMetrics([createSrDataItem({ sr_type: 'linstor' })])
-
-    assert.equal(metrics.length, 3)
-    for (const m of metrics) {
-      assert.equal(m.labels.sr_type, 'linstor')
-    }
-  })
-
-  it('should omit sr_type label when empty', () => {
-    const metrics = formatSrMetrics([createSrDataItem({ sr_type: '' })])
-
-    assert.equal(metrics.length, 3)
-    for (const m of metrics) {
-      assert.equal(m.labels.sr_type, undefined)
-    }
-  })
-
-  it('should distinguish multiple SRs by sr_type label', () => {
-    const metrics = formatSrMetrics([
-      createSrDataItem({ uuid: 'sr-a', sr_type: 'linstor' }),
-      createSrDataItem({ uuid: 'sr-b', sr_type: 'lvm' }),
-    ])
-
-    const typesByUuid = new Map<string, string>()
-    for (const m of metrics) {
-      const srUuid = m.labels.sr_uuid
-      const srType = m.labels.sr_type
-      if (srUuid !== undefined && srType !== undefined) {
-        typesByUuid.set(srUuid, srType)
-      }
-    }
-    assert.equal(typesByUuid.get('sr-a'), 'linstor')
-    assert.equal(typesByUuid.get('sr-b'), 'lvm')
-  })
-
-  it('should round-trip sr_type through formatToOpenMetrics', () => {
-    const metrics = formatSrMetrics([createSrDataItem({ sr_type: 'linstor' })])
-    const output = formatToOpenMetrics(metrics)
-    assert.match(output, /sr_type="linstor"/)
-  })
-})
-
-// ============================================================================
-// VDI Metrics Tests
-// ============================================================================
-
-describe('formatVdiMetrics', () => {
-  const createVdiDataItem = (overrides: Partial<VdiDataItem> = {}): VdiDataItem => ({
-    uuid: 'vdi-uuid-123',
-    name_label: 'System Disk',
-    size: 107374182400, // 100 GiB
-    usage: 53687091200, // 50 GiB
-    sr_uuid: 'sr-uuid-456',
-    sr_name: 'Local Storage',
-    sr_type: 'lvm',
-    pool_id: 'pool-789',
-    pool_name: 'Production Pool',
-    ...overrides,
-  })
-
-  it('should generate 2 metrics per VDI with all labels when attached to a VM', () => {
-    const vdiData: VdiDataItem[] = [
-      createVdiDataItem({
-        vm_uuid: 'vm-uuid-abc',
-        vm_name: 'My VM',
-      }),
-    ]
-
-    const metrics = formatVdiMetrics(vdiData)
-
-    assert.equal(metrics.length, 2)
-
-    // Virtual size metric
-    const virtualSize = metrics[0]!
-    assert.equal(virtualSize.name, 'xcp_vdi_virtual_size_bytes')
-    assert.equal(virtualSize.type, 'gauge')
-    assert.equal(virtualSize.help, 'VDI virtual size in bytes')
-    assert.equal(virtualSize.value, 107374182400)
-    assert.equal(virtualSize.labels.pool_id, 'pool-789')
-    assert.equal(virtualSize.labels.pool_name, 'Production Pool')
-    assert.equal(virtualSize.labels.sr_uuid, 'sr-uuid-456')
-    assert.equal(virtualSize.labels.sr_name, 'Local Storage')
-    assert.equal(virtualSize.labels.vdi_uuid, 'vdi-uuid-123')
-    assert.equal(virtualSize.labels.vdi_name, 'System Disk')
-    assert.equal(virtualSize.labels.vm_uuid, 'vm-uuid-abc')
-    assert.equal(virtualSize.labels.vm_name, 'My VM')
-
-    // Physical usage metric
-    const physicalUsage = metrics[1]!
-    assert.equal(physicalUsage.name, 'xcp_vdi_physical_usage_bytes')
-    assert.equal(physicalUsage.type, 'gauge')
-    assert.equal(physicalUsage.help, 'VDI physical space used in bytes (allocated on SR)')
-    assert.equal(physicalUsage.value, 53687091200)
-    assert.equal(physicalUsage.labels.vm_uuid, 'vm-uuid-abc')
-    assert.equal(physicalUsage.labels.vm_name, 'My VM')
-  })
-
-  it('should omit vm labels when VDI is not attached to a VM', () => {
-    const vdiData: VdiDataItem[] = [createVdiDataItem()]
-
-    const metrics = formatVdiMetrics(vdiData)
-
-    assert.equal(metrics.length, 2)
-    assert.equal(metrics[0]!.labels.vm_uuid, undefined)
-    assert.equal(metrics[0]!.labels.vm_name, undefined)
-    assert.equal(metrics[1]!.labels.vm_uuid, undefined)
-    assert.equal(metrics[1]!.labels.vm_name, undefined)
-  })
-
-  it('should return empty array for empty input', () => {
-    const metrics = formatVdiMetrics([])
-
-    assert.equal(metrics.length, 0)
-  })
-
-  it('should generate correct number of metrics for multiple VDIs', () => {
-    const vdiData: VdiDataItem[] = [
-      createVdiDataItem({ uuid: 'vdi-1', name_label: 'Disk 1' }),
-      createVdiDataItem({ uuid: 'vdi-2', name_label: 'Disk 2' }),
-      createVdiDataItem({ uuid: 'vdi-3', name_label: 'Disk 3' }),
-    ]
-
-    const metrics = formatVdiMetrics(vdiData)
-
-    assert.equal(metrics.length, 6) // 2 metrics * 3 VDIs
-  })
-
-  it('should map size to virtual_size_bytes and usage to physical_usage_bytes', () => {
-    const vdiData: VdiDataItem[] = [
-      createVdiDataItem({
-        size: 1000,
-        usage: 500,
-      }),
-    ]
-
-    const metrics = formatVdiMetrics(vdiData)
-
-    const virtualSize = metrics.find(m => m.name === 'xcp_vdi_virtual_size_bytes')!
-    const physicalUsage = metrics.find(m => m.name === 'xcp_vdi_physical_usage_bytes')!
-
-    assert.equal(virtualSize.value, 1000)
-    assert.equal(physicalUsage.value, 500)
-  })
-
-  it('should omit pool_name label when pool_name is empty', () => {
-    const vdiData: VdiDataItem[] = [createVdiDataItem({ pool_name: '' })]
-
-    const metrics = formatVdiMetrics(vdiData)
-
-    assert.equal(metrics[0]!.labels.pool_name, undefined)
-  })
-
-  it('should propagate sr_type label when present', () => {
-    const vdiData: VdiDataItem[] = [createVdiDataItem({ sr_type: 'linstor' })]
-
-    const metrics = formatVdiMetrics(vdiData)
-
-    assert.equal(metrics.length, 2)
-    for (const m of metrics) {
-      assert.equal(m.labels.sr_type, 'linstor')
-    }
-  })
-
-  it('should omit sr_type label when empty', () => {
-    const vdiData: VdiDataItem[] = [createVdiDataItem({ sr_type: '' })]
-
-    const metrics = formatVdiMetrics(vdiData)
-
-    for (const m of metrics) {
-      assert.equal(m.labels.sr_type, undefined)
-    }
-  })
-
-  it('should produce valid OpenMetrics output', () => {
-    const vdiData: VdiDataItem[] = [
-      createVdiDataItem({
-        vm_uuid: 'vm-uuid-abc',
-        vm_name: 'My VM',
-      }),
-    ]
-
-    const metrics = formatVdiMetrics(vdiData)
-    const output = formatToOpenMetrics(metrics)
-
-    assert.ok(output.includes('xcp_vdi_virtual_size_bytes'))
-    assert.ok(output.includes('xcp_vdi_physical_usage_bytes'))
-    assert.ok(output.includes('vdi_uuid="vdi-uuid-123"'))
-    assert.ok(output.includes('vm_uuid="vm-uuid-abc"'))
-  })
-
-  it('should escape special characters in label values', () => {
-    const vdiData: VdiDataItem[] = [
-      createVdiDataItem({
-        name_label: 'Disk "with quotes"',
-      }),
-    ]
-
-    const metrics = formatVdiMetrics(vdiData)
-    const output = formatToOpenMetrics(metrics)
-
-    assert.ok(output.includes('vdi_name="Disk \\"with quotes\\""'))
-  })
-})
-
-// ============================================================================
-// formatVmStatusMetrics Tests
-// ============================================================================
-
-describe('formatVmStatusMetrics', () => {
-  it('should return empty array for empty input', () => {
-    const result = formatVmStatusMetrics([])
-    assert.deepEqual(result, [])
-  })
-
-  it('should create one metric per VM with value 1', () => {
-    const vms: VmStatusItem[] = [
-      {
-        uuid: 'vm-1',
-        name_label: 'VM 1',
-        power_state: 'Running',
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-      {
-        uuid: 'vm-2',
-        name_label: 'VM 2',
-        power_state: 'Halted',
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-    ]
-
-    const result = formatVmStatusMetrics(vms)
-
-    assert.equal(result.length, 2)
-    assert.equal(result[0]!.value, 1)
-    assert.equal(result[1]!.value, 1)
-  })
-
-  it('should set correct metric name and type', () => {
-    const vms: VmStatusItem[] = [
-      {
-        uuid: 'vm-1',
-        name_label: 'VM 1',
-        power_state: 'Running',
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-    ]
-
-    const result = formatVmStatusMetrics(vms)
-
-    assert.equal(result[0]!.name, 'xcp_vm_status')
-    assert.equal(result[0]!.type, 'gauge')
-  })
-
-  it('should include all expected labels', () => {
-    const vms: VmStatusItem[] = [
-      {
-        uuid: 'vm-1',
-        name_label: 'VM 1',
-        power_state: 'Running',
-        pool_id: 'pool-1',
-        pool_name: 'Production',
-      },
-    ]
-
-    const result = formatVmStatusMetrics(vms)
-    const labels = result[0]!.labels
-
-    assert.equal(labels.pool_id, 'pool-1')
-    assert.equal(labels.pool_name, 'Production')
-    assert.equal(labels.uuid, 'vm-1')
-    assert.equal(labels.vm_name, 'VM 1')
-    assert.equal(labels.power_state, 'Running')
-  })
-
-  it('should omit pool_name when empty', () => {
-    const vms: VmStatusItem[] = [
-      { uuid: 'vm-1', name_label: 'VM 1', power_state: 'Running', pool_id: 'pool-1', pool_name: '' },
-    ]
-
-    const result = formatVmStatusMetrics(vms)
-
-    assert.equal(result[0]!.labels.pool_name, undefined)
-  })
-
-  it('should handle all VM power_state values', () => {
-    const vms: VmStatusItem[] = [
-      { uuid: 'vm-0', name_label: 'VM 0', power_state: 'Running', pool_id: 'pool-1', pool_name: 'Pool' },
-      { uuid: 'vm-1', name_label: 'VM 1', power_state: 'Paused', pool_id: 'pool-1', pool_name: 'Pool' },
-      { uuid: 'vm-2', name_label: 'VM 2', power_state: 'Halted', pool_id: 'pool-1', pool_name: 'Pool' },
-      { uuid: 'vm-3', name_label: 'VM 3', power_state: 'Suspended', pool_id: 'pool-1', pool_name: 'Pool' },
-    ]
-
-    const result = formatVmStatusMetrics(vms)
-
-    assert.equal(result.length, 4)
-    assert.equal(result[0]!.labels.power_state, 'Running')
-    assert.equal(result[1]!.labels.power_state, 'Paused')
-    assert.equal(result[2]!.labels.power_state, 'Halted')
-    assert.equal(result[3]!.labels.power_state, 'Suspended')
-  })
-
-  it('should produce valid OpenMetrics output', () => {
-    const vms: VmStatusItem[] = [
-      {
-        uuid: 'vm-1',
-        name_label: 'VM 1',
-        power_state: 'Running',
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-      {
-        uuid: 'vm-2',
-        name_label: 'VM 2',
-        power_state: 'Halted',
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-    ]
-
-    const metrics = formatVmStatusMetrics(vms)
-    const output = formatToOpenMetrics(metrics)
-
-    assert.ok(
-      output.includes(
-        '# HELP xcp_vm_status VM power state indicator (always 1; current state is carried by the power_state label)'
-      )
-    )
-    assert.ok(output.includes('# TYPE xcp_vm_status gauge'))
-    assert.ok(output.includes('power_state="Running"'))
-    assert.ok(output.includes('power_state="Halted"'))
-
-    // HELP and TYPE should appear only once
-    const helpCount = (output.match(/# HELP xcp_vm_status/g) || []).length
-    const typeCount = (output.match(/# TYPE xcp_vm_status/g) || []).length
-    assert.equal(helpCount, 1)
-    assert.equal(typeCount, 1)
-  })
-
-  it('should escape special characters in VM names', () => {
-    const vms: VmStatusItem[] = [
-      {
-        uuid: 'vm-1',
-        name_label: 'VM "with quotes"',
-        power_state: 'Running',
-        pool_id: 'pool-1',
-        pool_name: 'Pool',
-      },
-    ]
-
-    const metrics = formatVmStatusMetrics(vms)
-    const output = formatToOpenMetrics(metrics)
-
-    assert.ok(output.includes('vm_name="VM \\"with quotes\\""'))
-  })
-
-  it('should omit vm_name when empty', () => {
-    const vms: VmStatusItem[] = [
-      { uuid: 'vm-1', name_label: '', power_state: 'Running', pool_id: 'pool-1', pool_name: 'Pool' },
-    ]
-
-    const result = formatVmStatusMetrics(vms)
-
-    assert.equal(result[0]!.labels.vm_name, undefined)
-  })
-})
-
-// ============================================================================
-// formatVmUptimeMetrics Tests
-// ============================================================================
-
-describe('formatVmUptimeMetrics', () => {
-  const createVmLabelContext = (
-    overrides: Partial<{
-      startTime: number | null
-      is_control_domain: boolean
-      name_label: string
-      pool_name: string
-      power_state: string
-    }> = {}
-  ): LabelContext => ({
-    hosts: [
-      {
-        hostId: 'host-1',
-        hostAddress: '192.168.1.1',
-        hostLabel: 'Host 1',
-        poolId: 'pool-456',
-        poolLabel: 'Production Pool',
-        sessionId: 'session-123',
-        protocol: 'https:',
-      },
-    ],
-    labels: {
-      vms: {
-        'vm-uuid-123': {
-          name_label: overrides.name_label ?? 'Test VM',
-          is_control_domain: overrides.is_control_domain ?? false,
-          vbdDeviceToVdiName: {},
-          vbdDeviceToVdiUuid: {},
-          vifIndexToNetworkName: {},
-          startTime: overrides.startTime !== undefined ? overrides.startTime : Math.floor(Date.now() / 1000) - 3600,
-          power_state: overrides.power_state ?? 'Running',
-          pool_id: 'pool-456',
-          pool_name: overrides.pool_name ?? 'Production Pool',
-        },
-      },
-      hosts: {
-        'host-1': {
-          name_label: 'Host 1',
-          pifDeviceToNetworkName: {},
-          startTime: Math.floor(Date.now() / 1000) - 7200,
-        },
-      },
-      srs: {},
-      srTruncatedToUuid: {},
-      vdiUuidToSrUuid: {},
-    },
-  })
-
-  it('should generate uptime metric for VM with valid startTime', () => {
-    const now = Math.floor(Date.now() / 1000)
-    const bootTime = now - 3600 // 1 hour ago
-    const labelContext = createVmLabelContext({ startTime: bootTime })
-
-    const metrics = formatVmUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 1)
-    const metric = metrics[0]!
-    assert.equal(metric.name, 'xcp_vm_uptime_seconds')
-    assert.equal(metric.type, 'gauge')
-    assert.equal(metric.help, 'VM uptime in seconds since boot')
-    assert.equal(metric.labels.pool_id, 'pool-456')
-    assert.equal(metric.labels.pool_name, 'Production Pool')
-    assert.equal(metric.labels.uuid, 'vm-uuid-123')
-    assert.equal(metric.labels.vm_name, 'Test VM')
-    // Value should be approximately 3600 (1 hour)
-    assert.ok(metric.value >= 3599 && metric.value <= 3601)
-  })
-
-  it('should skip VM with null startTime', () => {
-    const labelContext = createVmLabelContext({ startTime: null })
-
-    const metrics = formatVmUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 0)
-  })
-
-  it('should skip VM-controller (is_control_domain)', () => {
-    const labelContext = createVmLabelContext({ is_control_domain: true })
-
-    const metrics = formatVmUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 0)
-  })
-
-  it('should generate metrics for multiple VMs', () => {
-    const now = Math.floor(Date.now() / 1000)
-    const labelContext: LabelContext = {
-      hosts: [
-        {
-          hostId: 'host-1',
-          hostAddress: '192.168.1.1',
-          hostLabel: 'Host 1',
-          poolId: 'pool-1',
-          poolLabel: 'Pool A',
-          sessionId: 'session-1',
-          protocol: 'https:',
-        },
-      ],
-      labels: {
-        vms: {
-          'vm-1': {
-            name_label: 'VM 1',
-            is_control_domain: false,
-            vbdDeviceToVdiName: {},
-            vbdDeviceToVdiUuid: {},
-            vifIndexToNetworkName: {},
-            startTime: now - 7200, // 2 hours
-            power_state: 'Running',
-            pool_id: 'pool-1',
-            pool_name: 'Pool A',
-          },
-          'vm-2': {
-            name_label: 'VM 2',
-            is_control_domain: false,
-            vbdDeviceToVdiName: {},
-            vbdDeviceToVdiUuid: {},
-            vifIndexToNetworkName: {},
-            startTime: now - 1800, // 30 minutes
-            power_state: 'Running',
-            pool_id: 'pool-1',
-            pool_name: 'Pool A',
-          },
-        },
-        hosts: {
-          'host-1': {
-            name_label: 'Host 1',
-            pifDeviceToNetworkName: {},
-            startTime: now - 86400,
-          },
-        },
-        srs: {},
-        srTruncatedToUuid: {},
-        vdiUuidToSrUuid: {},
-      },
-    }
-
-    const metrics = formatVmUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 2)
-    const vm1Metric = metrics.find(m => m.labels.uuid === 'vm-1')
-    const vm2Metric = metrics.find(m => m.labels.uuid === 'vm-2')
-    assert.ok(vm1Metric)
-    assert.ok(vm2Metric)
-    assert.ok(vm1Metric.value >= 7199 && vm1Metric.value <= 7201)
-    assert.ok(vm2Metric.value >= 1799 && vm2Metric.value <= 1801)
-  })
-
-  it('should omit pool_name when empty', () => {
-    const labelContext = createVmLabelContext({ pool_name: '' })
-
-    const metrics = formatVmUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 1)
-    assert.equal(metrics[0]!.labels.pool_name, undefined)
-  })
-
-  it('should omit vm_name when empty', () => {
-    const labelContext = createVmLabelContext({ name_label: '' })
-
-    const metrics = formatVmUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 1)
-    assert.equal(metrics[0]!.labels.vm_name, undefined)
-  })
-
-  it('should skip halted VM with stale startTime', () => {
-    const now = Math.floor(Date.now() / 1000)
-    const labelContext = createVmLabelContext({ startTime: now - 86400, power_state: 'Halted' })
-
-    const metrics = formatVmUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 0)
-  })
-
-  it('should skip suspended VM', () => {
-    const now = Math.floor(Date.now() / 1000)
-    const labelContext = createVmLabelContext({ startTime: now - 3600, power_state: 'Suspended' })
-
-    const metrics = formatVmUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 0)
-  })
-
-  it('should skip paused VM', () => {
-    const now = Math.floor(Date.now() / 1000)
-    const labelContext = createVmLabelContext({ startTime: now - 3600, power_state: 'Paused' })
-
-    const metrics = formatVmUptimeMetrics(labelContext)
-
-    assert.equal(metrics.length, 0)
-  })
-})
+import type { XostorAlarmsItem, XostorClusterItem, XostorSmartHost, XostorUpdateItem } from './index.mjs'
 
 // ============================================================================
 // formatXostorClusterMetrics Tests
